+++ CHANGES.markdown
@@ -1,6 +1,12 @@
 ## next
   - Add --times to print times for individual spec items
   - Add location information for failing afterAll-hooks
+  - Renamed:
+    - `headerFormatter` to `formatterHeader`
+    - `exampleGroupStarted` to `formatterGroupStarted`
+    - `exampleGroupDone` to `formatterGroupDone`
+    - `exampleStarted` to `formatterItemStarted`
+    - `exampleProgress` to `formatterProgress`
   - Remove `--verbose` option (this has been a noop since at least 2013)
   - Remove `--out` option (use shell output redirection instead)